requirements.txt
cffi
Click
cryptography
pexpect
ptyprocess
pycparser
pyOpenSSL
PyYAML
six
asn1crypto
